:recycle: refactor : 스키마 및 타입 수정
