+++ src/apis/emotion-log/types.ts
@@ -1,0 +1,3 @@
+import { EMOTION_STATUS } from '@/constants/emotions';
+
+export type Emotion = (typeof EMOTION_STATUS)[number];
