+++ src/apis/epigram/types.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+import { epigramResponseSchema, tagSchema } from './schemas';
+
+export type Tag = z.infer<typeof tagSchema>;
+
+export type EpigramResponse = z.infer<typeof epigramResponseSchema>;
