+++ src/apis/user/types.ts
@@ -1,0 +1,4 @@
+import { z } from 'zod';
+import { userSchema } from './schemas';
+
+export type User = z.infer<typeof userSchema>;
