+++ src/components/auth/LoginForm.tsx
@@ -6,8 +6,8 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { loginSchema } from '@/schemas';
-import { LoginForm as LoginFormType } from '@/types';
+import { loginSchema } from '@/apis/auth/schemas';
+import { LoginForm as LoginFormType } from '@/apis/auth/types';
 import Input from '@/components/ui/Field/Input';
 import Button from '@/components/ui/buttons';
 import { useModalStore } from '@/stores/ModalStore';
