+++ src/components/auth/SignupForm.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import Image from 'next/image';
-import { signupSchema } from '@/schemas';
-import { SignupForm as SignupFormType } from '@/types';
+import { signupSchema } from '@/apis/auth/schemas';
+import { SignupForm as SignupFormType } from '@/apis/auth/types';
 import Input from '@/components/ui/Field/Input';
 import Button from '@/components/ui/buttons';
 import OpendEye from '@/assets/icons/opend_eye.svg';
